// src/types/idea.ts
export type Idea = {
  id: number;
  id_owner: string;
  title: string;
  kategori: string;
  investment_amount: string;
  summary: string;
  description: string;
  status: boolean;
  image: string;
  location: string;
  is_verified: boolean;
  investment_available: string;
  created_at: string;
  updated_at: string;
};